requirements.txt
streamlit==0.63.0
pandas==1.0.5
matplotlib==3.1.1
neatbio==0.0.2
numpy==1.17.2
biopython==1.78
scikit-learn==0.23.1
scipy==1.3.1
virtualenv==16.7.9